42288 - move AbstractNodeTest to functional
